Move modified.AxisUtils.js -> SkeletonUtil.js
